requirements.txt
Flask
Flask-SocketIO
eventlet
flask_cors
python-dotenv==1.0.1